Separate SIS API and scraper functions
SIS API functions have been moved to a separate file as I felt the previous scraper file was a bit too large and hard to navigate.

The API functions were designed to be modular and used in isolation anyway, so this was an easy change.
